backend/pyproject.toml
[project]
name = "ui-demo-backend"
version = "0.1.0"
description = "Backend for visual YAML automation platform"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.32",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "pyyaml>=6.0.2",
  "alembic>=1.13.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
